Relocate component files to proper folder containers and components
