+++ src/containers/auth/Registration.js
