+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './containers/App';
+import App from './components/App';
 import reportWebVitals from './reportWebVitals';
 
 import {Provider} from 'react-redux';
